+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * Breaking: `import_url` does not append `.rb` to your url anymore. - KrauseFx
 * Minor core documentation updates - orta
+* Fixes to the markdown support in `warn`, `fail` and `message` - orta
 
 ## 0.9.1